app-cli/requirements.txt
msal==1.34.0
plotly==6.3.1
pyinstaller==6.16.0
pyyaml==6.0.3
requests==2.32.5
ruff==0.13.3
typer==0.19.2
